+++ src/ThreeDAnalyzer.Web/obj/Debug/net10.0/ThreeDAnalyzer.Web.AssemblyInfo.cs
@@ -13,7 +13,7 @@ using System.Reflection;
 [assembly: System.Reflection.AssemblyCompanyAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyConfigurationAttribute("Debug")]
 [assembly: System.Reflection.AssemblyFileVersionAttribute("1.0.0.0")]
-[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+f33782b7ea21e758a3e29917d92b7480c86e335c")]
+[assembly: System.Reflection.AssemblyInformationalVersionAttribute("1.0.0+cf10c5defc79fabfc490df288a8076ce15fa51fd")]
 [assembly: System.Reflection.AssemblyProductAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyTitleAttribute("ThreeDAnalyzer.Web")]
 [assembly: System.Reflection.AssemblyVersionAttribute("1.0.0.0")]